Restructure EditorNodes
